clean temp files as well

+++ clean-install.sh
@@ -1,7 +1,21 @@
 #!/bin/bash
-
-echo "----[ cleaning yarn cache"
+echo "----[ cleaning yarn & npm cache"
 yarn cache clean
+npm cache clean
+
+# exit on error
+set -e
+echo "----[ removing temp dirs ./dll, ./release, ./out"
+rm -rf ./out
+rc=$?
+rm -rf ./dll
+rc2=$?
+rm -rf ./release
+rc3=$?
+if [ $rc != 0 ] || [ $rc2 != 0 ] || [ $rc3 != 0 ] ; then
+    echo "====[ Error: Failed to remove temp dirs"
+    exit $rc
+fi
 
 echo "----[ removing ./node_modules npm packages"
 rm -rf ./node_modules
